+++ src/entities/node/ui/CheckpointNode/CheckpointNode.tsx
@@ -1,19 +1,13 @@
 import DeleteOutlineIcon from '@mui/icons-material/DeleteOutline'
-import {
-	Button,
-	Dialog,
-	DialogActions,
-	DialogContent,
-	DialogContentText,
-	DialogTitle,
-	IconButton
-} from '@mui/material'
+import { IconButton } from '@mui/material'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { Handle, NodeProps, Position, useReactFlow } from '@xyflow/react'
 import { nanoid } from 'nanoid'
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { toast } from 'react-toastify'
+
+import { DeleteNodeConfirmDialog } from '../DeleteNodeConfirmDialog'
 
 import styles from './CheckpointNode.module.css'
 import {
@@ -149,36 +143,15 @@ export const CheckpointNode = ({ data, id }: NodeProps<CustomNode>) => {
 				/>
 			) : null}
 
-			<Dialog
-				open={confirmOpen}
+			<DeleteNodeConfirmDialog
+				isOpen={confirmOpen}
+				nodeName={nodeName || t('labels.withoutName', { ns: 'nodes' })}
 				onClose={() => setConfirmOpen(false)}
-			>
-				<DialogTitle>{t('dialogs.deleteTitle', { ns: 'nodes' })}</DialogTitle>
-				<DialogContent>
-					<DialogContentText>
-						{t('dialogs.deleteConfirm', { ns: 'nodes' })}{' '}
-						<b>{nodeName || t('labels.withoutName', { ns: 'nodes' })}</b>?
-					</DialogContentText>
-				</DialogContent>
-				<DialogActions>
-					<Button
-						onClick={() => setConfirmOpen(false)}
-						variant='contained'
-					>
-						{t('buttons.cancel', { ns: 'common' })}
-					</Button>
-					<Button
-						onClick={() => {
-							handleDelete()
-							setConfirmOpen(false)
-						}}
-						color='error'
-						variant='contained'
-					>
-						{t('buttons.delete', { ns: 'common' })}
-					</Button>
-				</DialogActions>
-			</Dialog>
+				onConfirm={() => {
+					handleDelete()
+					setConfirmOpen(false)
+				}}
+			/>
 		</div>
 	)
 }
